pyproject.toml
[project]
name = "monorepo"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "faststream[rabbit] (>=0.5.48,<0.6.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "aiobotocore (>=2.24.2,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
ruff = "^0.12.10"
pre-commit = "^4.3.0"
factory-boy = "^3.3.3"
async-factory-boy = "^1.0.1"

[tool.ruff]
line-length = 88
src = ["app", "user_service", "tests"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
combine-as-imports = true

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
